server/ai/requirements.txt
tensorflow>=2.20.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
